+++ src/components/icon/styles.scss
@@ -1,4 +1,4 @@
-@import '../../css/settings';
+@import 'settings';
 
 .c-icon {
   width: 20px;
