--- src/components/legend/legend-component.md
@@ -1,2 +1,0 @@
-```js
-<Legend />
