+++ src/components/legend/legend-list/styles.scss
@@ -1,0 +1,6 @@
+.c-legend-list {
+  margin: 0;
+  padding: 0;
+  width: 100%;
+  overflow: auto;
+}
