+++ src/components/legend/styles.scss
@@ -1,0 +1,209 @@
+@import 'settings';
+
+.c-legend-map {
+  position: relative;
+  width: 100%;
+  max-width: 400px;
+  background: $color-white;
+  z-index: 2;
+  border-radius: 4px;
+
+  .legend-title {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    margin: 0;
+    padding: 10px 15px;
+    text-transform: uppercase;
+    color: $color-dark-1;
+    font-weight: $font-weight-regular;
+    font-size: 11px;
+  }
+
+  .toggle-legend {
+    position: absolute;
+    right: 0px;
+    width: 40px;
+    height: 32px;
+    top: -28px;
+    border: 0;
+    border-radius: 2px 2px 0 0;
+    background-color: rgba($color-white, .2);
+    cursor: pointer;
+
+    &:focus { outline: none; }
+
+    svg {
+      margin-top: 4px;
+      fill: $color-white;
+    }
+  }
+
+  .open-legend {
+    display: flex;
+    flex-flow: column;
+    max-height: 300px;
+    height: 0;
+    overflow: hidden;
+    opacity: 0;
+    // Without an auto height, the legend looks a bit 💩
+    // but we can't animate the height if max-height is
+    // used
+    // I've prefered to remove the animation than having
+    // a legend too tall - Clément
+
+    &.-active {
+      height: auto;
+      opacity: 1;
+      overflow: auto;
+    }
+
+    // > .legend-list {
+    //   margin: 0;
+    //   padding: 0;
+    //   width: 100%;
+    //   overflow: auto;
+    // }
+  }
+
+  .close-legend {
+    display: none;
+
+    &.-active {
+      display: inherit;
+    }
+
+    .toggle-legend {
+      position: relative;
+      top: auto;
+      padding: 0;
+      width: auto;
+      height: auto;
+      padding: 0;
+      background-color: $color-white;
+
+      svg {
+        margin-top: 0;
+        fill: $color-dark-1;
+      }
+    }
+  }
+}
+
+.c-legend-unit {
+  position: relative;
+  z-index: 10;
+  list-style: none;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  padding: 14px 15px 14px 10px; // We reduce the left padding because of the handle
+  border-bottom: 1px solid $border-color-1;
+
+  &:last-child {
+    border-bottom: 0;
+  }
+
+  .legend-info {
+    flex-grow: 1;
+    padding-left: 20px;
+  }
+
+  .legend-item-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: flex-start;
+
+    > h3 {
+      font-weight: $font-weight-bold;
+      flex-grow: 1;
+      flex-shrink: 1;
+    }
+  }
+
+  .handler {
+    position: absolute;
+    cursor: move;
+    top: 13px;
+
+    svg { fill: $color-dark-1; }
+  }
+
+  h3 {
+    font-size: $font-size-default;
+    color: $color-dark-1;
+    margin: 0;
+    margin-bottom: 3px;
+
+    &.-dark { color: $color-dark-1; }
+
+    .category {
+      font-weight: 400;
+      text-transform: capitalize;
+      padding: 0 8px 0 0;
+    }
+  }
+
+  .item-actions {
+    margin-left: 20px; // Margin with the title
+    flex-shrink: 0;
+    display: flex;
+    align-items: center;
+
+    button {
+      display: inline-block; // Needed to solve height issues
+      margin-right: 12px;
+      padding: 0;
+      border: 0;
+      background-color: transparent;
+      cursor: pointer;
+
+      &.-disabled {
+        cursor: default;
+        opacity: .5;
+
+        &:hover .c-icon {
+          fill: $color-dark-1;
+        }
+      }
+
+      &:last-child {
+        margin-right: 0;
+      }
+
+      &:hover .c-icon {
+        fill: $color-secondary;
+      }
+
+      .c-icon {
+        fill: $color-dark-1;
+      }
+    }
+
+    .info {
+      .c-icon { width: 14px; }
+    }
+
+    .close {
+      .c-icon { width: 11px; }
+    }
+
+    .toggle {
+      .c-icon { width: 16px; }
+    }
+
+    .opacity {
+      .c-icon { width: 11px; }
+    }
+
+    .layers {
+      .c-icon { width: 15px; }
+    }
+  }
+
+  &.-sort {
+    background-color: rgba(255, 255, 255, .9);
+    border-radius: 4px;
+  }
+}
