+++ src/css/settings.scss
@@ -1,14 +1,23 @@
-$border-color-1: rgba(26, 28, 34, 0.1);
-$border-color-2: rgba(202, 204,208, 0.85);; // For inputs
-
+// Colors
 $color-primary: #c32d7b;
 $color-secondary: #2C75B0;
 $color-tertiary: #FAB72E;
+$color-white: #FFF;
+$color-black: #000;
+
+$color-dark-1: #393f44;
+$color-dark-2: #717171;
+$color-dark-3: #CACCD0;
+$color-dark-4: #F4F6F7;
+
+// Border
+$border-color-1: rgba(26, 28, 34, 0.1);
+$border-color-2: rgba(202, 204,208, 0.85);; // For inputs
 
 // Typography
 $font-size-extra-small: 12px;
 $font-size-small: 13px;
-$font-size-normal: 14px;
+$font-size-default: 14px;
 $font-size-medium: 16px;
 $font-size-big: 21px;
 
@@ -16,3 +25,8 @@ $font-size-h1: 64px;
 $font-size-h2: 42px;
 $font-size-h3: 26px;
 $font-size-h4: 16px;
+
+// Weights
+$font-weight-light: 300;
+$font-weight-regular: 400;
+$font-weight-bold: 700;
